ata-tidy-fixes-01(06) -bugprone-implicit-widening-of-multiplication-result, part 1.

+++ src/include/miopen/solver/implicitgemm_util.hpp
@@ -445,9 +445,9 @@ static inline size_t ComputeLDSRequiredSize(const ProblemDescription& problem,
 
     // Multiplied worst_case_alignment_adjustment by 2 as
     // Both A and B matrix LDS size is increased.
-    const std::size_t lds_size =
-        (BPerBlock + KPerBlock) * EPerBlock * EPACKSize * GetTypeSize(problem.in_data_type) * 2 +
-        2 * worst_case_alignment_adjustment;
+    const std::size_t lds_size = (static_cast<std::size_t>(BPerBlock) + KPerBlock) * EPerBlock *
+                                     EPACKSize * GetTypeSize(problem.in_data_type) * 2 +
+                                 2 * static_cast<std::size_t>(worst_case_alignment_adjustment);
 
     return lds_size;
 }
